refactor: 회원 탈퇴 기능 위치 변경 및 refresh token 제거 부분 리팩터링
