+++ src/main/java/com/chooz/auth/presentation/AuthController.java
@@ -17,6 +17,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.CookieValue;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -67,22 +68,21 @@ public class AuthController {
 
     @PostMapping("/sign-out")
     public ResponseEntity<Void> signOut(
-            @CookieValue(name = CustomHeader.CustomCookie.REFRESH_TOKEN, required = false) String refreshToken,
             HttpServletResponse response,
             @AuthenticationPrincipal UserInfo userInfo
     ) {
-        if (Objects.isNull(refreshToken)) {
-            throw new BadRequestException(ErrorCode.INVALID_REFRESH_TOKEN_HEADER);
-        }
         refreshTokenCookieGenerator.removeCookie(response);
-        authService.signOut(userInfo.userId(), refreshToken);
+        authService.signOut(userInfo.userId());
         return ResponseEntity.ok().build();
     }
 
-    @PostMapping("/withdraw")
+    @DeleteMapping("/withdraw")
     public ResponseEntity<Void> withdraw(
-            @AuthenticationPrincipal UserInfo userInfo
+            @AuthenticationPrincipal UserInfo userInfo,
+            HttpServletResponse response
     ) {
+        refreshTokenCookieGenerator.removeCookie(response);
+        authService.withdraw(userInfo.userId());
         return ResponseEntity.ok().build();
     }
 }
